+++ packages/isaacscript-common/src/features/customStage/v.ts
@@ -1,0 +1,3 @@
+import { CustomStageData } from "../../interfaces/CustomStageData";
+
+export const customStages = new Map<string, CustomStageData>();