+++ packages/isaacscript-common/src/interfaces/CustomStageData.ts
@@ -1,0 +1,6 @@
+import { LevelStage } from "isaac-typescript-definitions";
+
+export interface CustomStageData {
+  name: string;
+  baseStage: LevelStage;
+}