+++ packages/docs/docs/main/change-log.md
@@ -12,7 +12,8 @@ This page lists the changes to the IsaacScript framework.
 ## July 12th, 2022
 
 - Breaking:
-  - `addRoomClearCharge` now has an argument of `bigRoomDoubleCharge` instead of `ignoreBigRoomDoubleCharge`, so you will need to invert the boolean.
+  - `addRoomClearCharge` now has an argument of `bigRoomDoubleCharge` (instead of the old argument of `ignoreBigRoomDoubleCharge`), so you will need to invert the boolean.
+  - `arrayRemove` will now only remove the first matching element (instead of every matching element). Use `arrayRemoveAll` for that behavior instead.
 - All of the `spawn` helper functions now have the option to pass an RNG object instead of a seed.
 - `addRoomClearCharge` and `addRoomClearChargeToSlot` now take an optional argument of `playSoundEffect`.
 - Added the following helper functions:
@@ -24,6 +25,9 @@ This page lists the changes to the IsaacScript framework.
   - `getGridEntitiesExcept`
   - `spawnCustomGrid`
   - `removeCustomGrid`
+  - `addCharge`
+  - `arrayRemoveAll`
+  - `arrayRemoveAllInPlace`
 
 ## July 5th, 2022